fix: it didn't compile so i fixed it.

+++ include/usb/common.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 namespace USB {
 
   uint8_t calc_crc5(const uint8_t *data, uint16_t len_bits);
@@ -10,5 +12,5 @@ namespace USB {
    * Should not include PID
    */
   void crc5_append(uint8_t *data, uint16_t len_bits);
-  void crc16_append(uint16_t *data, uint16_t len_bits);
+  void crc16_append(uint8_t *data, uint16_t len_bits);
 }
